+++ crates/ironrdp-client/src/rdp.rs
@@ -3293,22 +3293,6 @@ mod tests {
         assert!(matches!(receiver.try_recv(), Ok(RdpInputEvent::Resize { .. })));
     }
 
-    #[test]
-    fn input_sender_reservation_prevents_state_changes_without_queue_capacity() {
-        let (sender, mut receiver) = RdpInputSender::channel(1);
-        let permit = sender.try_reserve().expect("the empty queue has capacity");
-        assert!(sender.try_reserve().is_err());
-
-        permit.send(RdpInputEvent::Resize {
-            width: 1024,
-            height: 768,
-            scale_factor: 100,
-            physical_size: None,
-        });
-
-        assert!(matches!(receiver.try_recv(), Ok(RdpInputEvent::Resize { .. })));
-    }
-
     #[cfg(feature = "rdpdr")]
     #[test]
     fn disabled_rdpdr_does_not_build_a_backend() {
