+++ crates/ironrdp-testsuite-extra/tests/main.rs
@@ -5,6 +5,8 @@ mod agent;
 mod async_framed;
 mod capture_helpers;
 mod client_config;
+mod client_input;
+mod client_rail;
 mod dvc_pipe_proxy;
 mod e2e;
 mod gateway_detect;
